+++ infrastructure/mysql/dao/kf_card.go
@@ -18,7 +18,7 @@ type KFCard struct {
 	Day           int                  `json:"day" gorm:"column:day" doc:"卡密的天数"`
 	ExpireTime    int64                `json:"expireTime" gorm:"column:expire_time" doc:"过期时间"`
 	LastLoginTime int64                `json:"lastLoginTime" gorm:"column:last_login_time" doc:"上次登录时间"`
-	Price         int64                `json:"price" gorm:"price" doc:"价钱,根据系统配置生成."`
+	Price         float64              `json:"price" gorm:"price" doc:"价钱,根据系统配置生成."`
 	Version       int                  `json:"version" gorm:"version" doc:"乐观锁"`
 }
